requirements.txt
numpy>=2.0.0
scikit-learn>=1.4.0
sqlalchemy==2.0.23
flask==3.0.0
python-dotenv==1.0.0
faker==20.1.0
schedule==1.2.1
colorama==0.4.6
tabulate==0.9.0
fpdf2==2.7.6
Werkzeug==3.0.1